+++ rs/earthbucks_lib/src/header_chain.rs
@@ -12,9 +12,9 @@ pub struct HeaderChain {
 }
 
 impl HeaderChain {
-    pub const LENGTH_TARGET_ADJ_PERIOD: u64 = Header::BLOCKS_PER_TARGET_ADJ_PERIOD;
-    pub const LENGTH_EXPIRY_PERIOD: u64 = Script::PKHXR_90D_60D_X_LOCK_REL as u64;
-    pub const LENGTH_SAFETY_PERIOD: u64 = HeaderChain::LENGTH_EXPIRY_PERIOD * 2;
+    pub const LENGTH_TARGET_ADJ_PERIOD: u32 = Header::BLOCKS_PER_TARGET_ADJ_PERIOD;
+    pub const LENGTH_EXPIRY_PERIOD: u32 = Script::PKHXR_90D_60D_X_LOCK_REL;
+    pub const LENGTH_SAFETY_PERIOD: u32 = HeaderChain::LENGTH_EXPIRY_PERIOD * 2;
 
     pub fn new() -> Self {
         Self {
